Fix up From<FFIArrowRecordBatch>, shift create_reader into common/fetch
