+++ src/arrow2/reader_async.rs
@@ -1,6 +1,6 @@
 use crate::arrow2::error::ParquetWasmError;
 use crate::arrow2::error::Result;
-use crate::common::fetch::{get_content_length, make_range_request};
+use crate::common::fetch::{get_content_length, create_reader};
 use arrow2::array::Array;
 use arrow2::chunk::Chunk;
 use arrow2::datatypes::Schema;
@@ -10,31 +10,7 @@ use arrow2::io::parquet::read::RowGroupMetaData;
 use arrow2::io::parquet::read::{read_columns_many_async, RowGroupDeserializer};
 use futures::future::BoxFuture;
 use parquet2::read::read_metadata_async as _read_metadata_async;
-use range_reader::{RangeOutput, RangedAsyncReader};
-
-/// Create a RangedAsyncReader
-fn create_reader(
-    url: String,
-    content_length: usize,
-    min_request_size: Option<usize>,
-) -> RangedAsyncReader {
-    // at least 4kb per s3 request. Adjust to your liking.
-    let min_request_size = min_request_size.unwrap_or(4 * 1024);
-
-    // Closure for making an individual HTTP range request to a file
-    let range_get = Box::new(move |start: u64, length: usize| {
-        let url = url.clone();
-
-        Box::pin(async move {
-            let data = make_range_request(url.clone(), start, length)
-                .await
-                .unwrap();
-            Ok(RangeOutput { start, data })
-        }) as BoxFuture<'static, std::io::Result<RangeOutput>>
-    });
-
-    RangedAsyncReader::new(content_length, min_request_size, range_get)
-}
+use range_reader::RangedAsyncReader;
 
 pub async fn read_metadata_async(
     url: String,
